+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 - Fix misaligned text in export file name field.
 - Fix warning about missing metafiles.
 - Fix warning about using absolute paths when overwriting an fbx file.
+- Fix exporting zero scale not importing correctly in 3ds Max.
+- Fix window size cutting off text in export window.
 
 ## [4.1.0-pre.2] - 2021-05-05
 ### Known Issues