+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Fix object name not being undone after undo convert to prefab.
 - Fix settings reset button not working.
 - Fix not being able to export to a new folder on Mac.
+- Fix hdrp materials being exported with max incandescence.
 
 ## [4.1.0-pre.2] - 2021-05-05
 ### Known Issues