+++ CHANGELOG.md
@@ -16,6 +16,9 @@
 - Fix settings reset button not working.
 - Fix not being able to export to a new folder on Mac.
 - Fix hdrp materials being exported with max incandescence.
+- Fix multiple materials not exporting properly.
+### Changed
+- Duplicate meshes no longer export as instances, but as multiple meshes. 
 
 ## [4.1.0-pre.2] - 2021-05-05
 ### Known Issues