+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Fix not being able to export to a new folder on Mac.
 - Fix hdrp materials being exported with max incandescence.
 - Fix warning about ExportSettings not being in a file called ExportSettings.cs.
+- Fix exception when exporting a missing or null mesh.
 
 ## [4.1.0-pre.2] - 2021-05-05
 ### Known Issues